Put the limits where limits live, and stop a node's claim from wrapping the hub's own gauge
Three things found by asking whether the limits listed in the pull request were
ones to ship with.

Two of them are. A node's visitors are first come, first served, so one name can
starve the other nineteen on that node; and at its bound a node refuses
well-behaved visitors alongside abusive ones. Both are real and both are the right
trade for now -- the first would need a scheduler nobody has measured starvation
to size, and the second is what refusing costs, which the node cannot avoid the
way the hub does because a visitor's TLS state is live for as long as the visitor.

But they were in a pull request body, which after a squash is a commit message
nobody reads six months later. They are §15 entries now, with the two things the
body did not say: that the first-come rule is not new with hub-side admission but
is now deliberate rather than an accident of where the allocation failed, and that
the denial this replaced was louder -- an OutOfMemoryError and a reconnect -- where
a full node now sits quiet and turns everyone away, at a cost to an attacker of
eight addresses, bounded by a MAX_PER_IP that was not written for this.

§9.3 ended with "nothing yet tells the hub what a node's bound is, so the hub keeps
sending visitors a full node will reset". This change is what makes that false and
it was left standing.

And a defect. Registry.visitorCapacity summed what the nodes claim into an int,
with nothing validating the claim -- deliberately, since a node that overstates
only stops the hub refusing on its behalf and then resets what it cannot take. The
sum is the one place a claim reaches past its own node, and two nodes at
Integer.MAX_VALUE wrapped it to -2 and would have put a negative capacity on the
hub's public page. It is a long now, asserted by a test that reads -2 without it.

+++ hub/src/main/java/io/jailscale/hub/HttpFront.java
@@ -261,7 +261,7 @@ final class HttpFront {
         // goes away, and why it reads 0 on a hub whose nodes are older than that field.
         row(b, "Visitors per name", SniRouter.MAX_PER_NAME + " at once, and no more than the node"
             + " serving it will hold");
-        int capacity = hub.registry().visitorCapacity();
+        long capacity = hub.registry().visitorCapacity();
         row(b, "Visitors the nodes will hold", (capacity > 0 ? capacity + " together" : "not advertised")
             + ", " + hub.router().visitorsInFlight() + " being served right now");
         // Both numbers, because either one alone misleads. The count is what admission checks; the
